% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{check_connection}
\alias{check_connection}
\title{Check if connecting to childes db is possible}
\usage{
check_connection(db_version = "current", db_args = NULL)
}
\arguments{
\item{db_version}{String of the name of database version to use}

\item{db_args}{List with host, user, and password defined}
}
\value{
Logical indicating whether a connection was successfully formed
}
\description{
Check if connecting to childes db is possible
}
