% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_collections}
\alias{get_collections}
\title{Get collections}
\usage{
get_collections(connection = NULL, db_version = "current", db_args = NULL)
}
\arguments{
\item{connection}{Deprecated, ignored (childesr now reads from the
childes-db dataset on Redivis)}

\item{db_version}{String of the name of database version to use}

\item{db_args}{Deprecated, ignored}
}
\value{
A `tbl` of Collection data
}
\description{
Get collections
}
\examples{
\dontrun{
get_collections()
}
}
