% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_contexts}
\alias{get_contexts}
\title{Get the utterances surrounding a token(s)}
\usage{
get_contexts(
  collection = NULL,
  language = NULL,
  corpus = NULL,
  role = NULL,
  role_exclude = NULL,
  age = NULL,
  sex = NULL,
  target_child = NULL,
  token,
  window = c(0, 0),
  remove_duplicates = TRUE,
  connection = NULL,
  db_version = "current",
  db_args = NULL
)
}
\arguments{
\item{collection}{A character vector of one or more names of collections}

\item{language}{A character vector of one or more languages}

\item{corpus}{A character vector of one or more names of corpora}

\item{role}{A character vector of one or more roles to include}

\item{role_exclude}{A character vector of one or more roles to exclude}

\item{age}{A numeric vector of an single age value or a min age value and max
age value (inclusive) in months. For a single age value, participants are
returned for which that age is within their age range; for two ages,
participants are returned for whose age overlaps with the interval between
those two ages.}

\item{sex}{A character vector of values "male" and/or "female"}

\item{target_child}{A character vector of one or more names of children}

\item{token}{A character vector of one or more token patterns (`\%` matches
any number of wildcard characters, `_` matches exactly one wildcard
character)}

\item{window}{A length 2 numeric vector of how many utterances before and
after each utterance containing the target token to retrieve}

\item{remove_duplicates}{A boolean indicating whether to remove duplicate
utterances from the results}

\item{connection}{Deprecated, ignored (childesr now reads from the
childes-db dataset on Redivis)}

\item{db_version}{String of the name of database version to use}

\item{db_args}{Deprecated, ignored}
}
\value{
A 'tbl' of Utterance data, filtered down by supplied arguments.
}
\description{
Get the utterances surrounding a token(s)
}
\examples{
\dontrun{
get_contexts(target_child = "Shem", token = "dog")
}
}
