% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_corpora}
\alias{get_corpora}
\title{Get corpora}
\usage{
get_corpora(connection = NULL, db_version = "current", db_args = NULL)
}
\arguments{
\item{connection}{Deprecated, ignored (childesr now reads from the
childes-db dataset on Redivis)}

\item{db_version}{String of the name of database version to use}

\item{db_args}{Deprecated, ignored}
}
\value{
A `tbl` of Corpus data
}
\description{
Get corpora
}
\examples{
\dontrun{
get_corpora()
}
}
