% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_database_version}
\alias{get_database_version}
\title{Get database version}
\usage{
get_database_version(connection = NULL)
}
\arguments{
\item{connection}{A connection to the CHILDES database}
}
\value{
The database version as a string
}
\description{
Get database version
}
\examples{
\dontrun{
get_database_version()
}
}
