% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_db_info}
\alias{get_db_info}
\title{Get information on database connection options}
\usage{
get_db_info()
}
\value{
List of database info, including `redivis_current` (the childes-db
  version that is the current Redivis release) and `redivis_versions` (a
  named vector mapping childes-db versions to Redivis dataset versions).
}
\description{
Retrieves hosted information about available versions of childes-db and
their corresponding versions of the childes-db dataset on Redivis. If the
hosted information cannot be retrieved or does not specify the Redivis
version mapping, a mapping built into the package is used instead.
}
\examples{
\dontrun{
get_db_info()
}
}
