% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_table}
\alias{get_table}
\title{Get table}
\usage{
get_table(connection, name)
}
\arguments{
\item{connection}{A connection to the CHILDES database}

\item{name}{String of a table name}
}
\value{
A `tbl`
}
\description{
Get table
}
