% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Get transcripts}
\usage{
get_transcripts(
  collection = NULL,
  corpus = NULL,
  target_child = NULL,
  connection = NULL,
  db_version = "current",
  db_args = NULL
)
}
\arguments{
\item{collection}{A character vector of one or more names of collections}

\item{corpus}{A character vector of one or more names of corpora}

\item{target_child}{A character vector of one or more names of children}

\item{connection}{Deprecated, ignored (childesr now reads from the
childes-db dataset on Redivis)}

\item{db_version}{String of the name of database version to use}

\item{db_args}{Deprecated, ignored}
}
\value{
A `tbl` of Transcript data, filtered down by supplied arguments
}
\description{
Get transcripts
}
\examples{
\dontrun{
get_transcripts()
}
}
