% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr-package.R
\docType{package}
\name{childesr-package}
\alias{childesr}
\alias{childesr-package}
\title{childesr: Accessing the 'CHILDES' Database}
\description{
Tools for connecting to 'CHILDES', an open repository for transcripts of parent-child interaction. For more information on the underlying data, see \url{https://langcog.github.io/childes-db-website/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/langcog/childesr}
  \item Report bugs at \url{https://github.com/langcog/childesr/issues}
}

}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

Authors:
\itemize{
  \item Alessandro Sanchez \email{amsan7@gmail.com} [contributor]
  \item Daniel Yurovsky \email{dyurovsky@gmail.com}
}

Other contributors:
\itemize{
  \item Kyle MacDonald \email{kylem412@gmail.com} [contributor]
  \item Stephan Meylan \email{meylan.stephan@gmail.com} [contributor]
  \item Jessica Mankewitz \email{jmankewitz@gmail.com} [contributor]
}

}
\keyword{internal}
