% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{connect_to_childes}
\alias{connect_to_childes}
\title{Connect to CHILDES}
\usage{
connect_to_childes(version = "current", connectionDetails = NULL)
}
\arguments{
\item{version}{String name of the version to use}

\item{version}{connectionDetails NULL or list with host, user, and password defined}
}
\value{
A DBIConnection object for the CHILDES database
}
\description{
Connect to CHILDES
}
\examples{
\dontrun{
con <- connect_to_childes(version="current", connectionDetails=NULL)
DBI::dbDisconnect(con)
}
}
