% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{connect_to_childes}
\alias{connect_to_childes}
\title{Connect to CHILDES}
\usage{
connect_to_childes()
}
\value{
A DBIConnection object for the CHILDES database
}
\description{
Connect to CHILDES
}
\examples{
\dontrun{
con <- connect_to_childes()
DBI::dbDisconnect(con)
}
}
