% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_collections}
\alias{get_collections}
\title{Get collections}
\usage{
get_collections(connection = NULL)
}
\arguments{
\item{connection}{A connection to the CHILDES database}
}
\value{
A `tbl` of Collection data. If `connection` is supplied, the result
  is collected locally, otherwise it remains remote.
}
\description{
Get collections
}
\examples{
\dontrun{
get_collections()
}
}
