% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_content}
\alias{get_content}
\title{Get content}
\usage{
get_content(content_type, collection = NULL, corpus = NULL, role = NULL,
  age = NULL, sex = NULL, child = NULL, connection)
}
\arguments{
\item{content_type}{One of "token" or "utterance"}

\item{collection}{A character vector of one or more names of collections}

\item{corpus}{A character vector of one or more names of corpora}

\item{role}{A character vector of one or more roles}

\item{age}{A numeric vector of an age or a min age (inclusive) and max age
(exclusive) in months}

\item{sex}{A character vector of values "male" and/or "female"}

\item{child}{A character vector of one or more names of children}

\item{connection}{A connection to the CHILDES database}
}
\description{
Get content
}
