% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_content}
\alias{get_content}
\title{Get content}
\usage{
get_content(content_type, collection = NULL, language = NULL,
  corpus = NULL, role = NULL, role_exclude = NULL, age = NULL,
  sex = NULL, child = NULL, token = NULL, connection = NULL,
  version = "current", connectionDetails = NULL)
}
\arguments{
\item{content_type}{One of "token" or "utterance"}

\item{collection}{A character vector of one or more names of collections}

\item{language}{A character vector of one or more languages}

\item{corpus}{A character vector of one or more names of corpora}

\item{role}{A character vector of one or more roles to include}

\item{role_exclude}{A character vector of one or more roles to exclude}

\item{age}{A numeric vector of an age or a min age (inclusive) and max age
(exclusive) in months}

\item{sex}{A character vector of values "male" and/or "female"}

\item{child}{A character vector of one or more names of children}

\item{token}{A character vector of one or more token patterns (`\%` matches
any number of wildcard characters, `_` matches exactly one wildcard
character)}

\item{connection}{A connection to the CHILDES database}

\item{version}{String with the version name (used as the database name)}

\item{connectionDetails}{List with with host, user, and password}
}
\description{
Get content
}
