% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_content}
\alias{get_content}
\title{Get content}
\usage{
get_content(
  content_type,
  collection = NULL,
  language = NULL,
  corpus = NULL,
  role = NULL,
  role_exclude = NULL,
  age = NULL,
  sex = NULL,
  target_child = NULL,
  token = NULL,
  stem = NULL,
  part_of_speech = NULL,
  connection
)
}
\arguments{
\item{content_type}{One of "token" or "utterance"}

\item{collection}{A character vector of one or more names of collections}

\item{language}{A character vector of one or more languages}

\item{corpus}{A character vector of one or more names of corpora}

\item{role}{A character vector of one or more roles to include}

\item{role_exclude}{A character vector of one or more roles to exclude}

\item{age}{A numeric vector of an single age value or a min age value and max
age value (inclusive) in months. For a single age value, participants are
returned for which that age is within their age range; for two ages,
participants are returned for whose age overlaps with the interval between
those two ages.}

\item{sex}{A character vector of values "male" and/or "female"}

\item{target_child}{A character vector of one or more names of children}

\item{token}{A character vector of one or more token patterns (`\%` matches
any number of wildcard characters, `_` matches exactly one wildcard
character)}

\item{stem}{A character vector of one or more stems}

\item{part_of_speech}{A character vector of one or more parts of speech}

\item{connection}{A connection to the CHILDES database}
}
\description{
Get content
}
