% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_corpora}
\alias{get_corpora}
\title{Get corpora}
\usage{
get_corpora(connection = NULL)
}
\arguments{
\item{connection}{A connection to the CHILDES database}
}
\value{
A `tbl` of Corpus data. If `connection` is supplied, the result is
  collected locally, otherwise it remains remote.
}
\description{
Get corpora
}
\examples{
\dontrun{
get_corpora()
}
}
