% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_corpora}
\alias{get_corpora}
\title{Get corpora}
\usage{
get_corpora(connection = NULL, db_version = "current", db_args = NULL)
}
\arguments{
\item{connection}{A connection to the CHILDES database}

\item{db_version}{String of the name of database version to use}

\item{db_args}{List with host, user, and password defined}
}
\value{
A `tbl` of Corpus data. If `connection` is supplied, the result
  remains a remote query, otherwise it is retrieved into a local tibble.
}
\description{
Get corpora
}
\examples{
\donttest{
get_corpora()
}
}
