% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_database_version}
\alias{get_database_version}
\title{Get database version}
\usage{
get_database_version(connection = NULL, db_version = "current",
  db_args = NULL)
}
\arguments{
\item{connection}{A connection to the CHILDES database}

\item{db_version}{String of the name of database version to use}

\item{db_args}{List with host, user, and password defined}
}
\value{
The database version as a string
}
\description{
Get database version
}
\examples{
\donttest{
get_database_version()
}
}
