% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_tokens}
\alias{get_tokens}
\title{Get tokens}
\usage{
get_tokens(collection = NULL, corpus = NULL, child = NULL, role = NULL,
  age = NULL, sex = NULL, token = NULL, pattern = NULL,
  connection = NULL)
}
\arguments{
\item{collection}{A character vector of one or more names of collections}

\item{corpus}{A character vector of one or more names of corpora}

\item{child}{A character vector of one or more names of children}

\item{role}{A character vector of one or more roles}

\item{age}{A numeric vector of an age or a min age and max age (in months)}

\item{sex}{A character vector of values "male" and/or "female"}

\item{token}{A character vector of one or more tokens}

\item{pattern}{A string with a pattern to use in simple pattern matching (`%`
matches any number of wildcard characters, `_` matches exactly one wildcard
character)}

\item{connection}{A connection to the CHILDES database}
}
\value{
A `tbl` of Token data, filtered down by collection, corpus, child,
  role, age, sex, and token supplied, if any.
}
\description{
Get tokens
}
\examples{
\dontrun{
get_tokens(token = "dog")
}
}
