% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Get transcripts}
\usage{
get_transcripts(collection = NULL, corpus = NULL, child = NULL,
  connection = NULL)
}
\arguments{
\item{collection}{A character vector of one or more names of collections}

\item{corpus}{A character vector of one or more names of corpora}

\item{child}{A character vector of one or more names of children}

\item{connection}{A connection to the CHILDES database}
}
\value{
A `tbl` of Transcript data, filtered down by collection, corpus, and
  child supplied, if any.
}
\description{
Get transcripts
}
\examples{
\dontrun{
get_transcripts()
}
}
