% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Get transcripts}
\usage{
get_transcripts(collection = NULL, corpus = NULL,
  target_child = NULL, connection = NULL, db_version = "current",
  db_args = NULL)
}
\arguments{
\item{collection}{A character vector of one or more names of collections}

\item{corpus}{A character vector of one or more names of corpora}

\item{target_child}{A character vector of one or more names of children}

\item{connection}{A connection to the CHILDES database}

\item{db_version}{String of the name of database version to use}

\item{db_args}{List with host, user, and password defined}
}
\value{
A `tbl` of Transcript data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get transcripts
}
\examples{
\donttest{
get_transcripts()
}
}
