% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_types}
\alias{get_types}
\title{Get types}
\usage{
get_types(collection = NULL, corpus = NULL, child = NULL, role = NULL,
  role_exclude = NULL, age = NULL, sex = NULL, type = NULL,
  connection = NULL)
}
\arguments{
\item{collection}{A character vector of one or more names of collections}

\item{corpus}{A character vector of one or more names of corpora}

\item{child}{A character vector of one or more names of children}

\item{role}{A character vector of one or more roles to include}

\item{role_exclude}{A character vector of one or more roles to exclude}

\item{age}{A numeric vector of an age or a min age (inclusive) and max age
(exclusive) in months}

\item{sex}{A character vector of values "male" and/or "female"}

\item{type}{A character vector of one or more type patterns (`%` matches
any number of wildcard characters, `_` matches exactly one wildcard
character)}

\item{connection}{A connection to the CHILDES database}
}
\value{
A `tbl` of Type data, filtered down by collection, corpus, child,
  role, age, sex, and token supplied, if any.
}
\description{
Get types
}
\examples{
\dontrun{
get_types()
}
}
