% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childesr.R
\name{get_utterances}
\alias{get_utterances}
\title{Get utterances}
\usage{
get_utterances(collection = NULL, corpus = NULL, role = NULL,
  age = NULL, sex = NULL, child = NULL, connection = NULL)
}
\arguments{
\item{collection}{A character vector of one or more names of collections}

\item{corpus}{A character vector of one or more names of corpora}

\item{role}{A character vector of one or more roles}

\item{age}{A numeric vector of an age or a min age and max age (in months)}

\item{sex}{A character vector of values "male" and/or "female"}

\item{child}{A character vector of one or more names of children}

\item{connection}{A connection to the CHILDES database}
}
\value{
A `tbl` of Utterance data, filtered down by collection, corpus,
  child, role, age, and sex supplied, if any.
}
\description{
Get utterances
}
\examples{
\dontrun{
get_utterances(child = "Shem")
}
}
